+++ api/src/lib.rs
@@ -13,7 +13,6 @@
 #[cfg(test)]
 #[macro_use]
 extern crate approx;
-extern crate failure;
 
 #[macro_use]
 extern crate failure_derive;