+++ api/CHANGELOG.md
@@ -12,6 +12,7 @@
 
 * [DEV-4385] `AsFraction`'s `Numerator` now returns `Option<Self::Numerator>` to
   handle the case of a per- unit Unit.
+* Switched to Rust edition 2018.
 
 ### Removed